refactor: moving getMessage logic so it can be forced in pdf generation
